Explicit finalizer
Example instances do not explicitly dispose subscriptions.  I believe that based on the way that _disposeMe is used, it should be called regardless of the state of isDisposing.  There should also be a ~MvxSubscriptionToken to call Dispose(false) - when the user lets the instance fall to the finalizer.  As-is, if the objects are cleaned up in the finalizer, the dispose action callback is not called and the subscription count and other items are left incorrect.
Note: Should check action is not null in constructor and throw argumentnullexception if it is?